refactor: move port binding parsing to lib.rs
